added component for FacetsLabel and media query for mobile/tablet
