+++ app/javascript/components/Facets/Facets.css
@@ -26,20 +26,6 @@
   border-bottom: 1px solid #71767A;
 }
 
-.serp-facets-wrapper .filter-heading{
-  margin-top: 0;
-  padding-left: 0.35rem;
-}
-
-.serp-facets-wrapper .filter-heading-label{
-  color: #71767A;
-  font-size: 1rem;
-  font-weight: 500;
-  margin-left: 5px;
-  position: absolute;
-  top: 4px;
-}
-
 .serp-facets-wrapper .usa-tag{
   background: #F0F0F0;
   color: #005EA2;
