+++ app/javascript/components/Facets/FacetsLabel.css
@@ -1,0 +1,13 @@
+.filter-heading{
+  margin-top: 0;
+  padding-left: 0.35rem;
+}
+
+.filter-heading-label{
+  color: #71767A;
+  font-size: 1rem;
+  font-weight: 500;
+  margin-left: 5px;
+  position: absolute;
+  top: 4px;
+}
